requirements.txt
PyGithub
selenium